@@ -88,13 +88,12 @@ setup(
     packages=find_packages(exclude=["tests", "examples"]),
     package_data={"macrokit": ["py.typed"]},
     ext_modules=ext_modules,
-    python_requires=">=3.7",
+    python_requires=">=3.9",
     classifiers=[
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
     ],
 )
